elasticsearch/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
click==8.1.7
distro==1.9.0
elastic-transport==8.15.0
elasticsearch==8.15.1
fastapi==0.115.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
openai==1.50.2
pydantic @ git+https://github.com/pydantic/pydantic@d587553df8628a935979157e47d19b2149ffcaaf
pydantic_core==2.24.0
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.38.6
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
